Build flat dependency trees while auditing Gradle projects (#976)
* Upgraded to the new gradle-dep-tree version, to make the scan faster and more memory-efficient.
